requirements.txt
attrs==23.1.0
blinker==1.7.0
Bootstrap-Flask==2.3.2
boto3==1.34.19
botocore==1.34.19
certifi==2023.7.22
click==8.1.7
dominate==2.8.0
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
Flask-WTF==1.2.1
greenlet==3.0.1
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
Mako==1.3.0
MarkupSafe==2.1.3
mccabe==0.7.0
numpy==1.26.2
opencv-python==4.8.1.78
outcome==1.3.0.post0
pandas==2.1.3
Pillow==10.1.0
pycodestyle==2.11.1
pyflakes==3.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzbar==0.1.9
s3transfer==0.10.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sourcery==1.14.0
SQLAlchemy==2.0.23
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
visitor==0.1.3
Werkzeug==3.0.1
wsproto==1.2.0
WTForms==3.1.1
